+++ src/main/java/icu/samnyan/aqua/sega/aimedb/handler/impl/Lookup2Handler.java
@@ -48,7 +48,7 @@ public class Lookup2Handler implements BaseHandler {
         long aimeId = -1;
         Optional<Card> card = cardRepository.findByLuid((String) requestMap.get("luid"));
         if (card.isPresent()) {
-            aimeId = card.get().getExtId().longValue();
+            aimeId = card.get().getExtId();
         }
 
         Map<String, Object> resultMap = new HashMap<>();
